Moved compact/row layout button to title row, and started adding responsiveness to the JobList headers (filter dropdown, and search)
